requirements.txt
Django==5.2.5
django-environ==0.12.0
django-filter==25.1
django-phonenumber-field==8.1.0
django-select2==8.4.1
django-tables2==2.7.5
Faker==37.6.0
phonenumbers==9.0.13
pillow==11.3.0
